refactor: renamed DataSoreImplementation to DataStore
