+++ src/apis/ds/DataStoreInterface.java
@@ -1,0 +1,13 @@
+package apis.ds;
+
+public interface DataStoreInterface {
+
+    ParseInputFileResponse parseInputFile(ParseInputFileRequest parseInputFileRequest);
+
+    WriteListToFileResponse writeListToFile(WriteListToFileRequest writeListToFileRequest);
+
+    WriteIntegerToFileResponse writeIntegerToFile(WriteIntegerToFileRequest writeIntegerToFileRequest);
+
+    StartPrototypeResponse startPrototype(StartPrototypeRequest startPrototypeRequest);
+
+}
