+++ src/apis/ds/DataStorePrototype.java
@@ -1,16 +1,16 @@
 package apis.ds;
 
 public class DataStorePrototype {
-    public static void prototypeCode(DataStore dataStore) {
+    public static void prototypeCode(DataStoreInterface dataStoreInterface) {
 
 //        start this process
-        StartPrototypeResponse startPrototype = dataStore.startPrototype(new StartPrototypeRequest());
+        StartPrototypeResponse startPrototype = dataStoreInterface.startPrototype(new StartPrototypeRequest());
 
 //        parse the input file into raw integers
-        ParseInputFileResponse parseInputFile = dataStore.parseInputFile(new ParseInputFileRequest(null, null, null));
+        ParseInputFileResponse parseInputFile = dataStoreInterface.parseInputFile(new ParseInputFileRequest(null, null, null));
 
 //        write those integers into the output file
-        WriteListToFileResponse writeListToFileResponse = dataStore.writeListToFile(new WriteListToFileRequest(null, null));
+        WriteListToFileResponse writeListToFileResponse = dataStoreInterface.writeListToFile(new WriteListToFileRequest(null, null));
 
     }
 }
